# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Leave Management API",
        "description": "A comprehensive Leave Management API with role-based access control for managing employee leave requests, approvals, and leave balance tracking.",
        "contact": {
            "email": "support@leavemanagement.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/leave-balance": {
            "get": {
                "tags": [
                    "Leave Balance"
                ],
                "summary": "Get leave balance",
                "description": "Retrieve the leave balance for the authenticated user, including total entitlement, used days, and remaining balance. HR and Admin can check other users' balance by providing user_id parameter.",
                "operationId": "b40dc33ddf224d9708a365b42c3a5b5b",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User ID to check balance for (HR/Admin only)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leave balance retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "annual_leave_entitlement": {
                                                    "type": "number",
                                                    "example": 30
                                                },
                                                "used_days": {
                                                    "type": "number",
                                                    "example": 12.5
                                                },
                                                "remaining_days": {
                                                    "type": "number",
                                                    "example": 17.5
                                                },
                                                "pending_requests_days": {
                                                    "description": "Days in pending leave requests",
                                                    "type": "number",
                                                    "example": 2
                                                },
                                                "breakdown": {
                                                    "properties": {
                                                        "approved_leaves": {
                                                            "description": "Number of approved leave requests",
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "pending_leaves": {
                                                            "description": "Number of pending leave requests",
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "rejected_leaves": {
                                                            "description": "Number of rejected leave requests",
                                                            "type": "integer",
                                                            "example": 0
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - General users can only view their own balance"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/leave-balance/all": {
            "get": {
                "tags": [
                    "Leave Balance"
                ],
                "summary": "Get all users' leave balance summary",
                "description": "Retrieve leave balance summary for all users. Only accessible by HR and Admin roles.",
                "operationId": "b5329cfc1bbeca19552acf2b4a518948",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filter by user role",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "general",
                                "hr",
                                "admin"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leave balance summary retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "user_role": {
                                                        "type": "string",
                                                        "example": "general"
                                                    },
                                                    "annual_leave_entitlement": {
                                                        "type": "number",
                                                        "example": 30
                                                    },
                                                    "used_days": {
                                                        "type": "number",
                                                        "example": 12.5
                                                    },
                                                    "remaining_days": {
                                                        "type": "number",
                                                        "example": 17.5
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - Only HR and Admin can access this endpoint"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}